Mve custom hooks to hooks directory
