+++ components/RuleDisputes/RuleDisputes.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useState, useEffect } from 'react';
 import Comments from '../Comments/Comments.js'
-// import { useLocalStorage } from '../../hooks/hooks.js';
+import { useAllDisputeComments } from '../../hooks/hooks.js';
 
 const RuleDisputeForm = () => {
   const [expansion, setExpansion] = useState('');
@@ -85,37 +85,7 @@ const RuleDisputeForm = () => {
   )
 }
 
-function useLocalStorage(key, initialValue) {
-  const [storedValue, setStoredValue] = useState(() => {
-    try {
-      const item = window ? window.localStorage.getItem(key) : null
-      return item ? JSON.parse(item) : initialValue
-    } catch (error) {
-      console.log(error)
-      return initialValue
-    }
-  })
-  return [storedValue, setStoredValue]
-}
 
-function useAllDisputeComments(key) {
-  const [ storedValue, setStoredValue ] = useLocalStorage('allDisputes')
-
-  const setValue = value => {
-    try {
-      setStoredValue(value)
-      let itemToSet = [value]
-      if (storedValue) {
-        const parsedStoredValue = JSON.parse(window.localStorage.getItem('allDisputes'))
-        itemToSet = [...parsedStoredValue, value]
-      }
-      window.localStorage.setItem(key, JSON.stringify(itemToSet));  
-    } catch (error) {
-      console.log(error)
-    }
-  }
-  return [storedValue, setValue]
-}
 
 export default RuleDisputeForm;
